user_service/requirements.txt
annotated-types==0.7.0
bcrypt==4.3.0
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
greenlet==3.2.1
grpcio==1.71.0
grpcio-tools==1.71.0
idna==3.10
jose==1.0.0
optional==0.0.1
passlib==1.7.4
protobuf==5.29.4
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
setuptools==79.0.1
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
